BOL-375 refactor: game 관련 model(in data module) package 분리
