+++ data/src/main/java/com/yapp/bol/data/datasource/game/GameDataSource.kt
@@ -1,11 +1,10 @@
 package com.yapp.bol.data.datasource.game
 
-import com.yapp.bol.data.model.group.response.GameApiResponse
+import com.yapp.bol.data.model.game.GameApiResponse
 import com.yapp.bol.domain.model.ApiResult
 import kotlinx.coroutines.flow.Flow
 
 interface GameDataSource {
 
     fun getGameList(groupId: Int): Flow<ApiResult<GameApiResponse>>
-
 }
